+++ lib/grandpa/grandpa_test.go
@@ -51,272 +51,6 @@ func newTestVoters(t *testing.T) []*Voter {
 	return voters
 }
 
-func TestCheckForEquivocation_NoEquivocation(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-	state.AddBlocksToState(t, st.Block, 3)
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-
-	vote := NewVoteFromHeader(h)
-	require.NoError(t, err)
-
-	for _, v := range voters {
-		equivocated := gs.checkForEquivocation(v, vote)
-		require.False(t, equivocated)
-	}
-}
-
-func TestCheckForEquivocation_WithEquivocation(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-
-	var branches []*types.Header
-	for {
-		_, branches = state.AddBlocksToState(t, st.Block, 3)
-		if len(branches) != 0 {
-			break
-		}
-	}
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-
-	vote := NewVoteFromHeader(h)
-	require.NoError(t, err)
-
-	voter := voters[0]
-
-	gs.votes[voter.key.AsBytes()] = vote
-
-	vote2 := NewVoteFromHeader(branches[0])
-	require.NoError(t, err)
-
-	equivocated := gs.checkForEquivocation(voter, vote2)
-	require.True(t, equivocated)
-
-	require.Equal(t, 0, len(gs.votes))
-	require.Equal(t, 1, len(gs.equivocations))
-	require.Equal(t, 2, len(gs.equivocations[voter.key.AsBytes()]))
-}
-
-func TestCheckForEquivocation_WithExistingEquivocation(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-
-	var branches []*types.Header
-	for {
-		_, branches = state.AddBlocksToState(t, st.Block, 8)
-		if len(branches) > 1 {
-			break
-		}
-	}
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-
-	vote := NewVoteFromHeader(h)
-	require.NoError(t, err)
-
-	voter := voters[0]
-
-	gs.votes[voter.key.AsBytes()] = vote
-
-	vote2 := NewVoteFromHeader(branches[0])
-	require.NoError(t, err)
-
-	equivocated := gs.checkForEquivocation(voter, vote2)
-	require.True(t, equivocated)
-
-	require.Equal(t, 0, len(gs.votes))
-	require.Equal(t, 1, len(gs.equivocations))
-
-	vote3 := NewVoteFromHeader(branches[1])
-	require.NoError(t, err)
-
-	equivocated = gs.checkForEquivocation(voter, vote3)
-	require.True(t, equivocated)
-
-	require.Equal(t, 0, len(gs.votes))
-	require.Equal(t, 1, len(gs.equivocations))
-	require.Equal(t, 3, len(gs.equivocations[voter.key.AsBytes()]))
-}
-
-func TestValidateMessage_Valid(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-	state.AddBlocksToState(t, st.Block, 3)
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-
-	kr, err := keystore.NewEd25519Keyring()
-	require.NoError(t, err)
-
-	msg, err := gs.CreateVoteMessage(h, kr.Alice)
-	require.NoError(t, err)
-
-	vote, err := gs.ValidateMessage(msg)
-	require.NoError(t, err)
-	require.Equal(t, h.Hash(), vote.hash)
-}
-
-func TestValidateMessage_InvalidSignature(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-	state.AddBlocksToState(t, st.Block, 3)
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-
-	kr, err := keystore.NewEd25519Keyring()
-	require.NoError(t, err)
-
-	msg, err := gs.CreateVoteMessage(h, kr.Alice)
-	require.NoError(t, err)
-
-	msg.message.signature[63] = 0
-
-	_, err = gs.ValidateMessage(msg)
-	require.Equal(t, err, ErrInvalidSignature)
-}
-
-func TestValidateMessage_SetIDMismatch(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-	state.AddBlocksToState(t, st.Block, 3)
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-
-	kr, err := keystore.NewEd25519Keyring()
-	require.NoError(t, err)
-
-	msg, err := gs.CreateVoteMessage(h, kr.Alice)
-	require.NoError(t, err)
-
-	gs.state.setID = 1
-
-	_, err = gs.ValidateMessage(msg)
-	require.Equal(t, err, ErrSetIDMismatch)
-}
-
-func TestValidateMessage_Equivocation(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-
-	var branches []*types.Header
-	for {
-		_, branches = state.AddBlocksToState(t, st.Block, 8)
-		if len(branches) != 0 {
-			break
-		}
-	}
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-
-	vote := NewVoteFromHeader(h)
-	require.NoError(t, err)
-
-	voter := voters[0]
-
-	gs.votes[voter.key.AsBytes()] = vote
-
-	kr, err := keystore.NewEd25519Keyring()
-	require.NoError(t, err)
-
-	msg, err := gs.CreateVoteMessage(branches[0], kr.Alice)
-	require.NoError(t, err)
-
-	_, err = gs.ValidateMessage(msg)
-	require.Equal(t, ErrEquivocation, err, gs.votes)
-}
-
-func TestValidateMessage_BlockDoesNotExist(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-	state.AddBlocksToState(t, st.Block, 3)
-
-	kr, err := keystore.NewEd25519Keyring()
-	require.NoError(t, err)
-
-	fake := &types.Header{
-		Number: big.NewInt(77),
-	}
-
-	msg, err := gs.CreateVoteMessage(fake, kr.Alice)
-	require.NoError(t, err)
-
-	_, err = gs.ValidateMessage(msg)
-	require.Equal(t, err, ErrBlockDoesNotExist)
-}
-
-func TestValidateMessage_IsNotDescendant(t *testing.T) {
-	st := newTestState(t)
-	voters := newTestVoters(t)
-
-	gs, err := NewService(st.Block, voters)
-	require.NoError(t, err)
-
-	var branches []*types.Header
-	for {
-		_, branches = state.AddBlocksToState(t, st.Block, 8)
-		if len(branches) != 0 {
-			break
-		}
-	}
-
-	h, err := st.Block.BestBlockHeader()
-	require.NoError(t, err)
-	gs.head = h.Hash()
-
-	kr, err := keystore.NewEd25519Keyring()
-	require.NoError(t, err)
-
-	msg, err := gs.CreateVoteMessage(branches[0], kr.Alice)
-	require.NoError(t, err)
-
-	_, err = gs.ValidateMessage(msg)
-	require.Equal(t, ErrDescendantNotFound, err, gs.votes)
-}
-
-func TestPubkeyToVoter(t *testing.T) {
-	voters := newTestVoters(t)
-	kr, err := keystore.NewEd25519Keyring()
-	require.NoError(t, err)
-
-	state := NewState(voters, 0, 0)
-	voter, err := state.pubkeyToVoter(kr.Alice.Public().(*ed25519.PublicKey))
-	require.NoError(t, err)
-	require.Equal(t, voters[0], voter)
-}
-
 func TestGetDirectVotes(t *testing.T) {
 	st := newTestState(t)
 	voters := newTestVoters(t)
@@ -361,18 +95,20 @@ func TestGetVotesForBlock_NoDescendantVotes(t *testing.T) {
 	gs, err := NewService(st.Block, voters)
 	require.NoError(t, err)
 
-	var branches []*types.Header
-	var chain []*types.Header
-
+	var leaves []common.Hash
 	for {
-		chain, branches = state.AddBlocksToState(t, st.Block, 8)
-		if len(branches) != 0 {
+		state.AddBlocksToState(t, st.Block, 8)
+		leaves = gs.blockState.Leaves()
+		if len(leaves) > 1 {
 			break
 		}
 	}
 
-	voteA := NewVoteFromHeader(chain[7])
-	voteB := NewVoteFromHeader(branches[0])
+	// 1/3 of voters equivocate; ie. vote for both blocks
+	voteA, err := NewVoteFromHash(leaves[0], st.Block)
+	require.NoError(t, err)
+	voteB, err := NewVoteFromHash(leaves[1], st.Block)
+	require.NoError(t, err)
 
 	for i, k := range kr.Keys {
 		voter := k.Public().(*ed25519.PublicKey).AsBytes()
@@ -551,8 +287,6 @@ func TestGetPossiblePreVotedPredecessors_VaryingPredecessor(t *testing.T) {
 	expectedAt7, err := common.HexToHash("0x57508d4d2c5b01e6bd50dacee5d14979a6f23e41d4b4eb6464a8a29015549847")
 	require.NoError(t, err)
 
-	t.Log(blocks[expectedAt6])
-
 	// this should return the highest common predecessor of (a, b) and (b, c) with >=2/3 votes,
 	// which are the nodes at depth 6 and 7.
 	require.Equal(t, 2, len(blocks))
@@ -619,10 +353,6 @@ func TestGetPossiblePreVotedPredecessors_VaryingPredecessor_MoreBranches(t *test
 	expectedAt7, err := common.HexToHash("0x57508d4d2c5b01e6bd50dacee5d14979a6f23e41d4b4eb6464a8a29015549847")
 	require.NoError(t, err)
 
-	for h := range blocks {
-		t.Log(h)
-	}
-
 	// this should return the highest common predecessor of (a, b) and (b, c) with >=2/3 votes,
 	// which are the nodes at depth 6 and 7.
 	require.Equal(t, 2, len(blocks))
@@ -639,18 +369,19 @@ func TestGetPossiblePreVotedBlocks_OneBlock(t *testing.T) {
 	gs, err := NewService(st.Block, voters)
 	require.NoError(t, err)
 
-	var branches []*types.Header
-	var chain []*types.Header
-
+	var leaves []common.Hash
 	for {
-		chain, branches = state.AddBlocksToState(t, st.Block, 8)
-		if len(branches) != 0 {
+		state.AddBlocksToState(t, st.Block, 8)
+		leaves = gs.blockState.Leaves()
+		if len(leaves) > 1 {
 			break
 		}
 	}
 
-	voteA := NewVoteFromHeader(chain[7])
-	voteB := NewVoteFromHeader(branches[0])
+	voteA, err := NewVoteFromHash(leaves[0], st.Block)
+	require.NoError(t, err)
+	voteB, err := NewVoteFromHash(leaves[1], st.Block)
+	require.NoError(t, err)
 
 	for i, k := range kr.Keys {
 		voter := k.Public().(*ed25519.PublicKey).AsBytes()
